% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-docs.R
\docType{data}
\name{ATC_Tree_UpperBound_2024}
\alias{ATC_Tree_UpperBound_2024}
\title{ATC Tree Upper Bound 2014}
\format{
A data frame 4 variables:
\describe{
  \item{ATCCode}{The code of ATC nodes}
  \item{Name}{The name of ATC nodes}
  \item{ATC_length}{The number of character in the ATCCode}
  \item{upperBound}{The index of the last son of the current node in the tree}
}
}
\source{
World Health Organization, ATC classification register
}
\usage{
ATC_Tree_UpperBound_2024
}
\description{
The expected form of the ATC tree, this version of the ATC tree is from the WHO
website (2024-02-23).
}
\keyword{datasets}
