% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ATCtoNumeric}
\alias{ATCtoNumeric}
\title{Convert ATC Code for each patients to the corresponding DFS number of the ATC tree}
\usage{
ATCtoNumeric(patientATC, tree)
}
\arguments{
\item{patientATC}{: patients observations, for each patient we got a string 
containing taken medications (ATC code)}

\item{tree}{: ATC tree (we assume that there is a column 'ATCCode' )}
}
\value{
a matrix of the same size as patientATC but containing integer 
that are the index of the corresponding ATC code.
}
\description{
Convert ATC Code for each patients to the corresponding DFS number of the ATC tree
}
\examples{
 ATC_code <- c('A01AA30 A01AB03', 'A10AC30')
 ATCtoNumeric(ATC_code, ATC_Tree_UpperBound_2024)

}
