\name{emcAdr-package}
\alias{emcAdr-package}
\alias{emcAdr}
\docType{package}
\title{
\packageTitle{emcAdr}
}
\description{
\packageDescription{emcAdr}
}
\author{
\packageAuthor{emcAdr}

Maintainer: \packageMaintainer{emcAdr}
}
\keyword{ package }
\seealso{
\code{\link[Rcpp:Rcpp-package]{Rcpp}}, 
\link[RcppArmadillo:RcppArmadillo]{RcppArmadillo}
}
