% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{FAERS_myopathy}
\alias{FAERS_myopathy}
\title{FAERS Myopathy Dataset}
\format{
A data frame with 2 columns:
\describe{
  \item{patientATC}{Drug intake for each patient as a vector of ATC tree indices}
  \item{patientADR}{Indicates if the patient experienced myopathy as an adverse event}
}
}
\source{
Food & Drug Administration Event Reporting System (FAERS)
}
\usage{
FAERS_myopathy
}
\description{
Example dataset representing drug intake and adverse event reports from FAERS.
This dataset is provided to demonstrate the functionality of genetic and MCMC algorithms in the package.
}
\keyword{datasets}
