% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{frequencyHist}
\alias{frequencyHist}
\title{test data transformation for histogram plot (in)}
\usage{
frequencyHist(RRDistribution, f)
}
\arguments{
\item{RRDistribution}{: the RR distribution Vector (given by the EMC algorithm)}

\item{f}{: the function used to plot, hist by default}
}
\description{
test data transformation for histogram plot (in)
}
