% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_dissimilarity_from_csv_file}
\alias{get_dissimilarity_from_csv_file}
\title{Recover the square matrix of distance between cocktails where the index (i,j)
of the matrix is the distance between cocktails i and j in the csv file containing
results of genetic algorithm}
\usage{
get_dissimilarity_from_csv_file(filename, ATCtree, normalization = TRUE)
}
\arguments{
\item{filename}{: the name of the file returned by the print_csv function.}

\item{ATCtree}{: ATC tree with upper bound of the DFS (without the root)}

\item{normalization}{: Do we keep the distance between cocktail in the range [0;1] ?}
}
\value{
The square matrix of distances between cocktails
}
\description{
Recover the square matrix of distance between cocktails where the index (i,j)
of the matrix is the distance between cocktails i and j in the csv file containing
results of genetic algorithm
}
