% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{histogramToDitribution}
\alias{histogramToDitribution}
\title{Convert the histogram returned by the DistributionApproximation function, to a real number ditribution
(that can be used in a test for example)}
\usage{
histogramToDitribution(vec)
}
\arguments{
\item{tree}{: ATC tree (we assume that there is a column 'ATCCode' )}

\item{patients}{: patients observations, for each patient we got a string containing every medication he takes/took}
}
\description{
Convert the histogram returned by the DistributionApproximation function, to a real number ditribution
(that can be used in a test for example)
}
