% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_cocktail_to_string_cocktail}
\alias{int_cocktail_to_string_cocktail}
\title{Function used to convert integer cocktails (like the one outputed by the distributionApproximation function)
to string cocktail in order to make them more readable}
\usage{
int_cocktail_to_string_cocktail(cocktails, ATC_name)
}
\arguments{
\item{cocktails}{cocktails vector to be converted (index in the ATC tree)}

\item{ATC_name}{The ATC_name column of the ATC tree}
}
\value{
The name of integer cocktails in cocktails
}
\description{
Function used to convert integer cocktails (like the one outputed by the distributionApproximation function)
to string cocktail in order to make them more readable
}
\examples{
\donttest{
  data("ATC_Tree_UpperBound_2024")
  int_list = list(c(561, 904),
               c(1902, 4585))
  int_cocktail_to_string_cocktail(int_list, ATC_Tree_UpperBound_2024$Name)
}
}
