% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OustandingScoreToDistribution}
\alias{OustandingScoreToDistribution}
\title{Output the outstanding score (Outstanding_score) outputed by the MCMC algorithm
in a special format}
\usage{
OustandingScoreToDistribution(outstanding_score, max_score)
}
\arguments{
\item{outstanding_score}{: Outstanding_score outputed by MCMC algorithm to be converted
to the ScoreDistribution format}

\item{max_score}{: max_score parameter used during the MCMC algorithm}
}
\value{
outstanding_score in a format compatible with MCMC algorithm output
}
\description{
Output the outstanding score (Outstanding_score) outputed by the MCMC algorithm
in a special format
}
