% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalTests.R
\name{p_value_greater_than_empirical}
\alias{p_value_greater_than_empirical}
\title{Calculate p-value of sampled value being greater than empirical distribution}
\usage{
p_value_greater_than_empirical(
  empirical_distribution,
  sampled_value,
  isFiltered = F
)
}
\arguments{
\item{empirical_distribution}{A numeric vector of values representing the empirical distribution (return value of DistributionAproximation function)}

\item{sampled_value}{A numeric value representing the sampled value}

\item{isFiltered}{A boolean representing if we want to use the filtered distribution or the distribution as is (False by default)}
}
\value{
A numeric value representing the empirical p-value
}
\description{
Calculate p-value of sampled value being greater than empirical distribution
}
