% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_csv}
\alias{print_csv}
\title{Print every cocktails found during the genetic algorithm when used with the 
hyperparam_test_genetic_algorithm function. This enables to condense the solutions 
found in each files by collapsing similar cocktail in a single row by cocktail.}
\usage{
print_csv(
  input_filenames,
  observations,
  repetition,
  ATCtree,
  csv_filename = "solutions.csv"
)
}
\arguments{
\item{input_filenames}{: A List containing filename of hyperparam_test_genetic_algorithm output file}

\item{observations}{: observation of the AE based on the medications of each patients
(a DataFrame containing the medication on the first column and the ADR (boolean) on the second)
on which we want to compute the risk distribution}

\item{repetition}{: The parameter nb_test_desired used in the hyperparam test function}

\item{ATCtree}{: ATC tree with upper bound of the DFS (without the root)}

\item{csv_filename}{: Name of the output file, "solutions.csv" by default}
}
\description{
Print every cocktails found during the genetic algorithm when used with the 
hyperparam_test_genetic_algorithm function. This enables to condense the solutions 
found in each files by collapsing similar cocktail in a single row by cocktail.
}
