% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{qq_plot_output}
\alias{qq_plot_output}
\title{Make a Quantile-Quantile diagram from the output of the MCMC algorithm (DistributionAproximation)
and the algorithm that exhaustively calculates the distribution}
\usage{
qq_plot_output(estimated, true, filtered = F, color = "steelblue")
}
\arguments{
\item{estimated}{Outputed object of DistributionApproximation function}

\item{true}{Outputed object of either DistributionApproximation function or True distribution
computation function}

\item{filtered}{Make use of the classic distributuion estimation or of the filtred one
(number of patient taking the cocktail > beta)}

\item{color}{The color of the dashed line of the qq-plot}
}
\description{
Make a Quantile-Quantile diagram from the output of the MCMC algorithm (DistributionAproximation)
and the algorithm that exhaustively calculates the distribution
}
