% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trueDistributionSizeTwoCocktail}
\alias{trueDistributionSizeTwoCocktail}
\title{The true RR distribution of cocktail of size 2}
\usage{
trueDistributionSizeTwoCocktail(ATCtree, observations, beta)
}
\arguments{
\item{ATCtree}{: ATC tree with upper bound of the DFS (without the root)}
}
\value{
the RR distribution among size 2 cocktail
}
\description{
The true RR distribution of cocktail of size 2
}
