% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{ATC_Tree_UpperBound_2024}
\alias{ATC_Tree_UpperBound_2024}
\title{ATC Tree Upper Bound 2024}
\format{
A data frame with 4 variables:
\describe{
  \item{ATCCode}{The code of ATC nodes}
  \item{Name}{The name of ATC nodes}
  \item{ATC_length}{The number of characters in the ATCCode}
  \item{upperBound}{The index of the last child node in the tree}
}
}
\source{
World Health Organization, ATC classification register
}
\usage{
ATC_Tree_UpperBound_2024
}
\description{
Example dataset representing the ATC tree structure, sourced from the WHO website (2024-02-23).
This dataset is provided for demonstration and testing purposes with the package.
}
\keyword{datasets}
