% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ATCtoNumeric}
\alias{ATCtoNumeric}
\title{Convert ATC Code for each patients to the corresponding DFS number of the ATC tree}
\usage{
ATCtoNumeric(patients, tree)
}
\arguments{
\item{patients}{: patients observations, for each patient we got a string containing every medication he takes/took}

\item{tree}{: ATC tree (we assume that there is a column 'ATCCode' )}
}
\description{
Convert ATC Code for each patients to the corresponding DFS number of the ATC tree
}
