% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalTests.R
\name{calculate_divergence}
\alias{calculate_divergence}
\title{Calculate the divergence between 2 distributions (the true Distribution and the learned one)}
\usage{
calculate_divergence(
  empirical_distribution,
  true_distribution,
  method = "TV",
  Filtered = F
)
}
\arguments{
\item{empirical_distribution}{A numeric vector of values representing the empirical distribution (return value of DistributionAproximation function)}

\item{true_distribution}{A numeric vector of values representing the true distribution computed by the trueDistributionSizeTwoCocktail function}

\item{method}{A string, either "TV" or "KL" to respectively use the total variation distance or the Kullback-Leibler divergence. (default = "TV")}
}
\value{
A numeric value representing the divergence of the 2 distributions
}
\description{
Calculate the divergence between 2 distributions (the true Distribution and the learned one)
}
