% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMC}
\alias{EMC}
\title{The Evolutionary MCMC method that runs the random walk}
\usage{
EMC(
  n,
  ATCtree,
  observations,
  nbIndividuals = 5L,
  nbResults = 5L,
  startingIndividuals = NULL,
  startingTemperatures = NULL
)
}
\arguments{
\item{n}{: number of step}

\item{ATCtree}{: ATC tree with upper bound of the DFS (without the root)}

\item{nbIndividuals}{: number on individuals in the population, 5 by default}

\item{nbResults}{: number of results returned (best RR individuals), 5 by default}

\item{startingIndividuals}{: starting individuals, randomly initialized by default 
(a DataFrame containing the medication on the first column and the ADR (boolean) on the second)}

\item{startingTemperatures}{: starting temperatures, randomly initialized by default}

\item{observation}{: real observation of the ADR based on the medications of each real patients
(same form as the starting individuals)}
}
\value{
if no problem return an R List with : the distribution of RR we've met; bests individuals and the corresponding RRs. Otherwise the list is empty
}
\description{
The Evolutionary MCMC method that runs the random walk
}
