% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-docs.R
\docType{data}
\name{FAERS_myopathy}
\alias{FAERS_myopathy}
\title{FAERS myopathy}
\format{
A data frame with 2 columns:
\describe{
  \item{patientATC}{Drug intake for each patient in the form of vector of index of the ATC tree}
  \item{patientADR}{Does the patient experience the considered Adverse Event ? (here : myopathy)}
}
}
\source{
Food & Drug Administration Event Reporting System data
}
\usage{
FAERS_myopathy
}
\description{
The dataframe corresponding to patients drug intake from the FAERS dataset. It
shows expected input for the genetic algorithm as well as the MCMC algorithm.
}
\keyword{datasets}
