% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GeneticAlgorithm}
\alias{GeneticAlgorithm}
\title{Genetic algorithm, trying to reach the best cocktail (the one which maximize
the fitness function, Related Risk in our case)}
\usage{
GeneticAlgorithm(
  epochs,
  nbIndividuals,
  ATCtree,
  observations,
  p_crossover = 0.8,
  p_mutation = 0.01,
  nbElite = 0L,
  tournamentSize = 2L
)
}
\arguments{
\item{epochs}{: number of step}

\item{nbIndividuals}{: size of the popultation}

\item{ATCtree}{: ATC tree with upper bound of the DFS (without the root)}

\item{nbElite}{: number of best individual we keep from generation to generation}

\item{tournamentSize}{: size of the tournament (select the best individual be tween tournamentSize individuals)}

\item{observation}{: real observation of the ADR based on the medications of each real patients
(a DataFrame containing the medication on the first column and the ADR (boolean) on the second)}

\item{p_crossover:}{probability to operate a crossover on the crossover phase.}

\item{p_mutation:}{probability to operate a mutation after the crossover phase.}
}
\value{
if no problem return the best cocktail we found (according to the fitness function which is the Relative Risk)
}
\description{
Genetic algorithm, trying to reach the best cocktail (the one which maximize
the fitness function, Related Risk in our case)
}
