% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{histogramToDitribution}
\alias{histogramToDitribution}
\title{Convert the histogram returned by the DistributionApproximation function, to a real number distribution
(that can be used in a test for example)}
\usage{
histogramToDitribution(vec)
}
\arguments{
\item{vec}{: distribution returned by the DistributionAproximationFunction}
}
\value{
A vector containing sampled risk during the MCMC algorithm
}
\description{
Convert the histogram returned by the DistributionApproximation function, to a real number distribution
(that can be used in a test for example)
}
\examples{
\dontrun{
  DistributionApproximationResults = DistributionApproximation(epochs = 1000, ...)
  histogramToDitribution(DistributionApproximationResults$Distribution)
}
}
