% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_value_csv_file}
\alias{p_value_csv_file}
\title{Used to add the p_value to each cocktail of a csv_file that is an
output of the genetic algorithm}
\usage{
p_value_csv_file(
  distribution_outputs,
  filename,
  filtred_distribution = TRUE,
  sep = ";"
)
}
\arguments{
\item{distribution_outputs}{A list of distribution of cocktails of different sizes
in order to compute the p_value for multiple cocktail sizes}

\item{filename}{The file name of the .csv file containing the output}

\item{sep}{The separator used in the csv file (';' by default)}

\item{filtered_distribution}{Does the p-values have to be computed using filtered distribution
or normal distribution (filtered distribution by default)}
}
\description{
Used to add the p_value to each cocktail of a csv_file that is an
output of the genetic algorithm
}
