% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_evolution}
\alias{plot_evolution}
\title{Plot the evolution of the mean and the best value of the population used by the GeneticAlgorithm}
\usage{
plot_evolution(
  list,
  mean_color = "#F2A900",
  best_color = "#008080",
  xlab = "Epochs",
  ylab = "Score"
)
}
\arguments{
\item{list}{A list with 2 elements returned by the GeneticAlgorithm: "mean" and "best", containing the numeric vectors representing the mean and best fitness of the population}

\item{mean_color}{A string specifying the color of the mean values}

\item{best_color}{A string specifying the color of the best values}

\item{xlab}{A string specifying the label for the x-axis}

\item{ylab}{A string specifying the label for the y-axis}
}
\description{
Plot the evolution of the mean and the best value of the population used by the GeneticAlgorithm
}
